+++ CHANGELOG.md
@@ -81,6 +81,8 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 * Update Roofline binaries
   * Rebuild using latest ROCm stack
   * OS distribution support minimum for roofline feature is now Ubuntu22.04, RHEL9, and SLES15SP6
+* Improve analysis block based filtering to accept metric id level filtering
+  * This can be used to collect individual metrics from various sections of analysis config
 
 ### Optimized